requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
django-redis==6.0.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
djangorestframework-stubs==3.16.8
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
GDAL==3.10.3
idna==3.11
inflection==0.5.1
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.2.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
utils==1.0.2
vine==5.1.0
wcwidth==0.6.0
